로그인 시 헤더에 표시할 정보 응답에 넣어서 보내도록 구현

+++ src/main/java/com/clover/salad/common/file/repository/FileUploadRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.clover.salad.common.file.entity.FileUploadEntity;
 
-public interface FileUploadRepository extends JpaRepository<FileUploadEntity, Long> {
+public interface FileUploadRepository extends JpaRepository<FileUploadEntity, Integer> {
 }
